Extracted a loader
and renamed the other loaders...
