+++ app/src/main/java/com/island/island/Activities/FeedActivity.java
@@ -8,12 +8,10 @@ import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.DialogFragment;
 import android.support.v4.app.Fragment;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -21,7 +19,7 @@ import android.view.ViewGroup;
 import com.island.island.Adapters.PostAdapter;
 import com.island.island.Database.IslndContract;
 import com.island.island.PostCollection;
-import com.island.island.PostLoader;
+import com.island.island.loader.LocalPostLoader;
 import com.island.island.R;
 import com.island.island.SimpleDividerItemDecoration;
 
@@ -51,7 +49,7 @@ public class FeedActivity extends Fragment {
         mRecyclerView.setLayoutManager(mLayoutManager);
 
         mAdapter = new PostAdapter(getContext(), null);
-        getLoaderManager().initLoader(0, null, new PostLoader(getContext(), mAdapter));
+        getLoaderManager().initLoader(0, null, new LocalPostLoader(getContext(), mAdapter));
 
         mRecyclerView.setAdapter(mAdapter);
         mRecyclerView.addItemDecoration(new SimpleDividerItemDecoration(getContext()));
